Disable CMD "Quick Edit Mode"
Fix a crashed when typing into a CMD window which had text selected.
Disabling the "Quick Edit Mode" makes it harder to block stdout by
accidentally clicking on a random spot in the console window.
To fix it completely no calls write calls to stdout should happen in the hook.