New strategy for low-level keyboard hook remapping
Do not run the remapping logic from within the low-level keyboard hook.
This caused a lot of problems with re-entrancy when sending inputs.
If a key is unmapped we do no longer forward it in the hook, instead we
enqueue all captured key events for later processing. The actual
processing now happens from an async task. We must ensure to re-send
also the unmapped keys from the processing taks.
The new approach has the downside that we catch every key event and
downstream keyboards hooks only ever see injected events. So far this
did not cause any issues in practice.